Add fast-quit to ft `diff` and `rebase`
When staging a file with git I tend to use it with the `-p` flag a lot.
In this view we end up in a `diff` filetype mode and I want to be able
to quickly quit out of it if I want to.

Same goes for when I view the list of commits in `gitrebase` filetype
mode. I want to quickly be able to get out of it.