Alias to kill coreaudiod
Because most of osx sucks and so upon waking laptop from sleep it causes
a bug with the audio to happen where nothing is being outputted. Killing
the core process for audio seems to do the trick, though most of the
time in the case audio stops working in Chrome you'd have to close that
down and re-open it.