Enable vim to persist undo between sessions
Now as I have moved from using Tabs to be using Buffers instead I
noticed that, when switching between buffers, I would lose my undo
history.

This commit resolves that, hwoever keep in mind that it requires me to
manually create the undodir for it to work as Vim will not do it for me.