Add Gutentags - tags manager plugin
Since I started working with tags in Vim and I wouldn't want to update
the tags manually everytime, I figured this plugin would do the trick
for me and be smart enough to do so in an efficient way without being in
my way.

So I'll see how it plays out.