Set `fzf` default opts and using `bat`
Setting default options that are passed to `fzf` and is also making use
of a new program `bat` which is `cat` on magic.

A lot of things are going on here but basically we define our preview
window with some if's or else's against the current selected file in the
list.

I also define bindings to toggle the preview window and to jump half a
page up/down in the file list outputted by `fzf`.