Ripgrep ignore .git/ directory
I thought I had given it a try without finding a solution on how to
ignore directories with `rg`. Today I gave it another try and found out
there is a flag `--glob` to do just that.. Silly I hadn't found this
earlier as it shows `.git/` object items when navigating through
projects in vim.