Add mapping to re-select last shape
This is super nifty as once I e.g. delete a block using "visual line"
and paste it to where I would want it it's often times not correctly
indented. Having to re-select it is bothersome and this mapping allows
me to quickly re-select last shape to directly allow me to correc the
indentation.