Add mapping to convert Postman JSON -> JS Object
When I'm using Postman I'm working with JSON, whenever I usually have
tested something or received a response I'd like to save in javascript
it's cumbersome having to turn JSON into a valid literal object by hand.
So I've created this mapping to do it for me, still has some caveats but
works most of the time.. what doesn't work is that it sometimes fails to
add the trailing comma to the last line inside of the object.