Add rtorrentrc to repo
I've moved over to start using rtorrent for my torrents and was met with
a lot of learnings due to its script language not being the "best" in
terms of being documented. Because of this I want to start persisting
what I have today and to save me from having to go through it again..
but by no means is it finished, this is merely my "checkpoint".