Add git utility function to read log commits
`fshow()` shows the history of commits and by pressing `ctrl-m` you
"open" the commit and have everything accessible to you as if you had
done `git show COMMIT_HASH`.

This is useful and should replace my current `ctrl-g ctrl-h` workflow
when I want to find the specific commit for a fixup or squash commit.`
However when pressing ENTER it doesn't give us the hash of the commit,
instead it opens it up as if you had pressed `ctrl-m`. This needs to get
fixed before it can replace my current workflow.