Clean up and organize .tmux.conf
Separating content with comments and grouping related options with each
other. Also for options I didn't remember what they meant, such as
`setw` I instead decided to skip their aliases and go with the full name
of it to be more descriptive and avoid having to search for what it
does.