Switch colorscheme to `ghdark`
When making the switch I realized I didn't have the correct colors as
seen in screenshots for the colorscheme.

Apparently the solution was the following[1]:
```
" Correct RGB escape codes for vim inside tmux
if !has('nvim') && $TERM ==# 'screen-256color'
  let &t_8f = "\<Esc>[38;2;%lu;%lu;%lum"
  let &t_8b = "\<Esc>[48;2;%lu;%lu;%lum"
endif
```

With the switch I also remove older colorschemes that are no longer used
and the colorscheme config was moved entirely to
`.vim/plugin/colorscheme.vim` instead.

I've also added `autocmd`'s for when I need to switch to a light theme
(colorscheme is `github`), do note I need to set the colorscheme twice
for the white background to show. Not sure why.

[1]: https://vi.stackexchange.com/a/13459