Add `tmux` WSL-only keybindings
This adds the possibility to set conditional settings in the
`.tmux.conf` file by sourcing the `~/.profile` file in order to get the
`determine_platform` function defined and set the value of it to
`TMUX_PLATFORM` to later be used for the conditions.

For this to work I had to have an escape hatch in the `.profile` file to
stop execution of the rest of the file since it otherwise broke the
command of `tmux setenv ...` as it expected a single word, not two lines
and not two words to be outputted... at the very least.