Correct impl. of platform-specific keybindings
In my previous commit:
7951083902580425f467afef45608fbbaf8147f1

The implementation was working for OSX but for some reason didn't on my
Windows laptop. Which was a good thing in the end since it wouldn't be
all that great having to source the `.profile`-file twice (once when the
shell is instantiated and once when the tmux session starts).

This other solution exports the `__OS_PLATFORM__` environment variable
which will from hereon be static and can be consumed by the `.tmux.conf`
to set platform-specific settings. Again better than sourcing `.profile`
twice.

Oh also updated a typo, it should've been `LINUX` as platform instead of
`WINDOWS` for the if-statement check.