Add OSX-style AutoHotKey mappings
This helps me to retain same keybindings as I have when I work on OSX.
Primarily the curly braces mapping was the main reason for me to seek
out a solution/workaround to get it to work the same way on Windows.

On Mac I've also (with Karabiner) set up so that Option+<h,j,k,l> would
move the cursor the same way as it would in Vim.  This was added too in
this commit to support that. Things _might_ break in applications that
I'm not aware of yet but I will take it as it comes.